+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-18 21:06:41
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-05-18 22:28:19
+// Commit: chore(i18n): update dependencies and cleanup unused imports in i18n
 
 class i18nComponentState {
-  final String commitRef = '1276';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-18T21:06:41+05:00');
+  final String commitRef = '1282';
+  final String lastMessage = 'chore(i18n): update dependencies and cleanup unused imports in i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-18T22:28:19+05:00');
 }